facttic/mv-issues#12 refactor and post tests

+++ post/dao.js
@@ -10,8 +10,8 @@ PostSchema.statics.createNew = async function createNew(post) {
   return newPost;
 };
 
-PostSchema.statics.insertMany = async function insertMany(posts) {
-  return await this.model("Post").collection.insertMany(posts, { ordered: false });
+PostSchema.statics.createMany = async function createMany(posts) {
+  return await PostDAO.createMany(posts, { ordered: false });
 };
 
 PostSchema.statics.getAll = async function getAll({ skip, limit, sort, query }) {
@@ -47,24 +47,23 @@ PostSchema.statics.getAllByManifestationId = async function getAllByManifestatio
   };
 };
 
+PostSchema.statics.getByPostIdStr = async function getByPostIdStr(postIdStr, source) {
+  return await PostDAO.findOne({ post_id_str: postIdStr, source }).exec();
+};
+
+PostSchema.statics.getById = async function getById(_id) {
+  return await PostDAO.findOne({ _id });
+};
+
 PostSchema.statics.removeByUserId = async function removeById(userIdStr, userId) {
   return await PostDAO.delete({ "user.id_str": userIdStr }, userId);
 };
 
-PostSchema.statics.countUsersByManifestationId = async function countUsersByManifestationId(
-  manifestationId,
-) {
-  const peopleCount = await PostDAO.distinct("user.id_str", {
-    manifestation_id: manifestationId,
-  }).exec();
-  return peopleCount.length;
-};
-
-PostSchema.statics.findByIdStr = async function findByIdStr(postIdStr, source) {
-  return await PostDAO.findOne({ post_id_str: postIdStr, source }).exec();
-};
-
 PostSchema.statics.removeById = async function removeById(_id, userId = null) {
+  const post = await PostDAO.findOne({ _id });
+  if (!post) {
+    throw new Error("Post does not exist");
+  }
   return await PostDAO.delete({ _id }, userId);
 };
 
@@ -74,6 +73,15 @@ PostSchema.statics.removeByManifestationId = async function removeByManifestatio
   userId = null,
 ) {
   return await PostDAO.delete({ _id, manifestation_id: manifestationId }, userId);
+};
+
+PostSchema.statics.countUsersByManifestationId = async function countUsersByManifestationId(
+  manifestationId,
+) {
+  const peopleCount = await PostDAO.distinct("user.id_str", {
+    manifestation_id: manifestationId,
+  }).exec();
+  return peopleCount.length;
 };
 
 PostSchema.plugin(mongooseDelete, {
